+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Minor fixes for Ac Device
+
 ## 2.12.2 (2023-09-06)
 
 * (theimo1221) Improve handling of inactive motion sensors